+++ src/gx/m2_lights.h
@@ -4,6 +4,7 @@
 #include <gfx/objects.h>
 
 #include <stdbool.h>
+#include <stddef.h>
 
 struct gx_m2_lights_init_data;
 struct gx_m2_instance;
@@ -19,8 +20,8 @@ struct gx_m2_lights
 
 void gx_m2_lights_init(struct gx_m2_lights *lights);
 void gx_m2_lights_destroy(struct gx_m2_lights *lights);
-bool gx_m2_lights_load(struct gx_m2_lights *lights, const struct wow_m2_light *wow_lights, uint32_t nb);
+bool gx_m2_lights_load(struct gx_m2_lights *lights, const struct wow_m2_light *wow_lights, size_t nb);
 void gx_m2_lights_initialize(struct gx_m2_lights *lights);
-void gx_m2_lights_render(struct gx_m2_lights *lights, const struct gx_m2_instance **instances, uint32_t nb);
+void gx_m2_lights_render(struct gx_m2_lights *lights, const struct gx_m2_instance **instances, size_t nb);
 
 #endif
